+++ src/renderer/src/icons/index.ts
@@ -1,0 +1,14 @@
+export { AnthropicIcon } from './AnthropicIcon';
+export { AzureIcon } from './AzureIcon';
+export { BedrockIcon } from './BedrockIcon';
+export { CerebrasIcon } from './CerebrasIcon';
+export { DeepseekIcon } from './DeepseekIcon';
+export { GeminiIcon } from './GeminiIcon';
+export { GroqIcon } from './GroqIcon';
+export { LmStudioIcon } from './LmStudioIcon';
+export { OllamaIcon } from './OllamaIcon';
+export { OpenAiIcon } from './OpenAiIcon';
+export { OpenAiCompatibleIcon } from './OpenAiCompatibleIcon';
+export { OpenRouterIcon } from './OpenRouterIcon';
+export { RequestyIcon } from './RequestyIcon';
+export { VertexAiIcon } from './VertexAiIcon';